LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-activation/activation-traverse-expected.txt



FAIL navigation.activation after traverse assert_equals: expected 2 but got 1
